requirements.txt
asgiref==3.6.0
coverage==7.2.5
Django==4.2.1
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
